refactor: rewriting ChainSandboxBody and moving it to a shared component
